+++ projects/np-ui-lib/src/lib/np-breadcrumb/np-breadcrumb.model.ts
@@ -2,7 +2,9 @@ export class NpBreadcrumbItem {
     constructor(item: any) {
         this.label = item.label;
         this.data = item.data;
+        this.disabled = item.disabled;
     }
     public label: string;
     public data: any;
+    public disabled: boolean;
 }