+++ projects/np-ui-lib/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## 10.3.3
 - Resolved data grid state variable reference issue
+- Add disabled property in breadcrumb model
 
 ## 10.3.2
 - Resolved client side filter issue for date in data grid